Move additional types to store.rs
I am starting to think it might be better to keep these types in
jam.rs, since I think the main value add of this work is to decouple
the store implementation from the parsing/execution implementation.
