+++ src/store.rs
@@ -1,4 +1,5 @@
 use std::slice::Iter;
+use thiserror::Error;
 
 use crate::{config::DesugaredTargetCfg, executor::ExecuteKind};
 
@@ -111,6 +112,66 @@ impl Clone for Shortcut {
         Shortcut(self.0.to_vec())
     }
 }
+
+#[derive(Debug, PartialEq)]
+pub enum Lookup {
+    NotFound,
+    Found,
+    ReconciliationFailure(String),
+    Conflict,
+}
+
+pub enum NextKey<'a> {
+    LeafKey { key: char, target_name: &'a str },
+    BranchKey { key: char },
+}
+
+impl<'a> NextKey<'a> {
+    pub fn new(key: char, target_names: Vec<&'a str>) -> NextKey<'a> {
+        if target_names.len() == 1 {
+            NextKey::LeafKey {
+                key,
+                target_name: target_names[0],
+            }
+        } else {
+            NextKey::BranchKey { key }
+        }
+    }
+
+    pub fn key(&self) -> char {
+        match self {
+            NextKey::LeafKey { key, .. } => *key,
+            NextKey::BranchKey { key } => *key,
+        }
+    }
+}
+
+#[derive(Error, Debug, PartialEq)]
+pub enum ExecError {
+    #[error("given shortcut '{shortcut}' is ambiguous (i.e. is it {conflict_msg}?)")]
+    Conflict {
+        shortcut: Shortcut,
+        conflict_msg: String,
+    },
+    #[error("no command for given shortcut '{shortcut}'")]
+    ShortcutNotFound { shortcut: Shortcut },
+    #[error("no command for given target name '{target_name}'")]
+    TargetNotFound { target_name: String },
+    #[error("target '{name}' has no executable function")]
+    CannotExec { name: String },
+    #[error("{description}")]
+    Reconciliation { description: String },
+    #[error("{description}")]
+    Executor { description: String },
+    #[error("failed to execute dependency ('{dep_name}'): {err}")]
+    Dependency {
+        dep_name: String,
+        err: Box<ExecError>,
+    },
+}
+
+// TODO: Maybe we just move all of this to jam.rs? The key value-add of the work is to just decouple the store implementation from the parsing/execution implementaiton (i.e. jam).
+pub type ExecResult<T> = Result<T, ExecError>;
 
 pub(crate) trait TargetStore {
     fn get(&self, key: &str) -> Option<&str>;
